add 'date of projects'; guest now can only read from database

+++ app/controllers/Proyek.php
@@ -36,20 +36,26 @@ class Proyek extends Controller
     {
 
         if (isset($_SESSION['ptpwl_user_id'])) {
-            if (empty($_POST['pt']) || empty($_POST['nama_proyek']) || empty($_POST['tahun_proyek']) || empty($_POST['lokasi']) || empty($_POST['maps']) || empty($_POST['link'])) {
-                Flasher::setFlash('gagal', ' ditambahkan', 'danger');
-                header('Location:' . BASEURL . 'proyek');
-                exit;
-            } else {
-                if ($this->model('Proyek_model')->tambahDataProyek($_POST) > 0) {
-                    Flasher::setFlash('berhasil', ' ditambahkan', 'success');
-                    header('Location:' . BASEURL . 'proyek');
-                    exit;
-                } else {
+            if ($_SESSION['ptpwl_user_id'] == 1) {
+                if (empty($_POST['pt']) || empty($_POST['nama_proyek']) || empty($_POST['tahun_proyek']) || empty($_POST['bulan_proyek']) || empty($_POST['tanggal_proyek']) || empty($_POST['lokasi']) || empty($_POST['maps']) || empty($_POST['link'])) {
                     Flasher::setFlash('gagal', ' ditambahkan', 'danger');
                     header('Location:' . BASEURL . 'proyek');
                     exit;
+                } else {
+                    if ($this->model('Proyek_model')->tambahDataProyek($_POST) > 0) {
+                        Flasher::setFlash('berhasil', ' ditambahkan', 'success');
+                        header('Location:' . BASEURL . 'proyek');
+                        exit;
+                    } else {
+                        Flasher::setFlash('gagal', ' ditambahkan', 'danger');
+                        header('Location:' . BASEURL . 'proyek');
+                        exit;
+                    }
                 }
+            } else {
+                Flasher::setFlash('anda tidak memiliki akses', ' ', 'danger');
+                header('Location:' . BASEURL . 'proyek');
+                exit;
             }
         } else {
             header('location:' . BASEURL . 'login');
@@ -58,12 +64,18 @@ class Proyek extends Controller
     public function hapus($id)
     {
         if (isset($_SESSION['ptpwl_user_id'])) {
-            if ($this->model('proyek_model')->hapusDataProyek($id) > 0) {
-                Flasher::setFlash('berhasil', ' dihapus', 'success');
-                header('Location:' . BASEURL . 'proyek');
-                exit;
+            if ($_SESSION['ptpwl_user_id'] == 1) {
+                if ($this->model('proyek_model')->hapusDataProyek($id) > 0) {
+                    Flasher::setFlash('berhasil', ' dihapus', 'success');
+                    header('Location:' . BASEURL . 'proyek');
+                    exit;
+                } else {
+                    Flasher::setFlash('gagal', ' dihapus', 'danger');
+                    header('Location:' . BASEURL . 'proyek');
+                    exit;
+                }
             } else {
-                Flasher::setFlash('gagal', ' dihapus', 'danger');
+                Flasher::setFlash('anda tidak memiliki akses', ' ', 'danger');
                 header('Location:' . BASEURL . 'proyek');
                 exit;
             }
@@ -79,27 +91,33 @@ class Proyek extends Controller
     public function ubah()
     {
         if (isset($_SESSION['ptpwl_user_id'])) {
-            if (empty($_POST['pt']) || empty($_POST['nama_proyek']) || empty($_POST['tahun_proyek']) || empty($_POST['lokasi']) || empty($_POST['maps']) || empty($_POST['link'])) {
-                Flasher::setFlash('gagal', ' diubah', 'danger');
-                header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
-                exit;
-            } else {
-                if ($this->model('Proyek_model')->ubahDataProyek($_POST) > 0) {
-                    Flasher::setFlash('berhasil', ' diubah', 'success');
-                    header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
-                    exit;
-                } else {
+            if ($_SESSION['ptpwl_user_id'] == 1) {
+                if (empty($_POST['pt']) || empty($_POST['nama_proyek']) || empty($_POST['tahun_proyek']) || empty($_POST['bulan_proyek']) || empty($_POST['tanggal_proyek']) || empty($_POST['lokasi']) || empty($_POST['maps']) || empty($_POST['link'])) {
                     Flasher::setFlash('gagal', ' diubah', 'danger');
                     header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
                     exit;
+                } else {
+                    if ($this->model('Proyek_model')->ubahDataProyek($_POST) > 0) {
+                        Flasher::setFlash('berhasil', ' diubah', 'success');
+                        header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
+                        exit;
+                    } else {
+                        Flasher::setFlash('gagal', ' diubah', 'danger');
+                        header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
+                        exit;
+                    }
                 }
+            } else {
+                Flasher::setFlash('anda tidak memiliki akses', ' ', 'danger');
+                header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
+                exit;
             }
         } else {
             header('location:' . BASEURL . 'login');
         }
     }
 
-    
+
     public function cari()
     {
         if (isset($_SESSION['ptpwl_user_id'])) {
